+++ Pod/Classes/TBSMEventHandler.h
@@ -7,10 +7,9 @@
 //
 
 #import <Foundation/Foundation.h>
-
 #import "TBSMTransition.h"
-
 #import "TBSMState.h"
+
 
 /**
  *  This class represents an event handler object.
@@ -26,7 +25,7 @@
 /**
  *  The target state of the transition triggered by the event.
  */
-@property (nonatomic, strong) TBSMState *target;
+@property (nonatomic, strong) id <TBSMTransitionVertex> target;
 
 /**
  *  The kind of transition to perform.
@@ -57,7 +56,7 @@
  *
  *  @return The event handler instance.
  */
-+ (instancetype)eventHandlerWithName:(NSString *)name target:(TBSMState *)target kind:(TBSMTransitionKind)kind action:(TBSMActionBlock)action guard:(TBSMGuardBlock)guard;
++ (instancetype)eventHandlerWithName:(NSString *)name target:(id <TBSMTransitionVertex>)target kind:(TBSMTransitionKind)kind action:(TBSMActionBlock)action guard:(TBSMGuardBlock)guard;
 
 /**
  *  Initializes a `TBSMEventHandler` from a given event name, target, action and guard.
@@ -72,6 +71,6 @@
  *
  *  @return An initialized `TBSMEventHandler` instance.
  */
-- (instancetype)initWithName:(NSString *)name target:(TBSMState *)target kind:(TBSMTransitionKind)kind action:(TBSMActionBlock)action guard:(TBSMGuardBlock)guard;
+- (instancetype)initWithName:(NSString *)name target:(id <TBSMTransitionVertex>)target kind:(TBSMTransitionKind)kind action:(TBSMActionBlock)action guard:(TBSMGuardBlock)guard;
 
 @end