+++ Pod/Classes/TBSMStateMachine.m
@@ -78,7 +78,7 @@
 - (void)setUp:(NSDictionary *)data
 {
     if (self.initialState) {
-        [self enterState:nil targetState:self.initialState data:data];
+        [self enter:nil targetState:self.initialState data:data];
     } else {
         @throw [NSException tb_noInitialStateException:self.name];
     }
@@ -87,7 +87,7 @@
 - (void)tearDown:(NSDictionary *)data
 {
     [self.scheduledEventsQueue cancelAllOperations];
-    [self exitState:self.currentState targetState:nil data:data];
+    [self exit:self.currentState targetState:nil data:data];
     _currentState = nil;
 }
 
@@ -96,7 +96,7 @@
 - (void)scheduleEvent:(TBSMEvent *)event
 {
     if (self.parentNode) {
-        TBSMStateMachine *topStateMachine = [self.parentNode parentNode];
+        TBSMStateMachine *topStateMachine = (TBSMStateMachine *)[self.parentNode parentNode];
         [topStateMachine scheduleEvent:event];
         return;
     }
@@ -116,8 +116,15 @@
         }
         NSArray *eventHandlers = [self.currentState eventHandlersForEvent:event];
         for (TBSMEventHandler *eventHandler in eventHandlers) {
-            TBSMTransition *transition = [TBSMTransition transitionWithSourceState:self.currentState targetState:eventHandler.target
-                                                                              kind:eventHandler.kind action:eventHandler.action guard:eventHandler.guard];
+            
+            TBSMTransition *transition = nil;
+            if ([eventHandler.target isKindOfClass:[TBSMState class]]) {
+                transition = [TBSMTransition transitionWithSourceState:self.currentState targetState:(TBSMState *)eventHandler.target
+                                                                  kind:eventHandler.kind action:eventHandler.action guard:eventHandler.guard];
+            } else {
+                transition = [TBSMCompoundTransition compoundTransitionWithSourceState:self.currentState targetPseudoState:(TBSMPseudoState *)eventHandler.target
+                                                                          action:eventHandler.action guard:eventHandler.guard];
+            }
             if ([transition performTransitionWithData:event.data]) {
                 return YES;
             }
@@ -134,10 +141,19 @@
     if (action) {
         action(sourceState, targetState, data);
     }
-    [self enterState:sourceState targetState:targetState data:data];
+    [self enter:sourceState targetState:targetState data:data];
 }
 
-- (void)enterState:(TBSMState *)sourceState targetState:(TBSMState *)targetState data:(NSDictionary *)data
+- (void)switchState:(TBSMState *)sourceState targetStates:(NSArray *)targetStates region:(TBSMParallelState *)region action:(TBSMActionBlock)action data:(NSDictionary *)data
+{
+    [self.currentState exit:sourceState targetState:region data:data];
+    if (action) {
+        action(sourceState, region, data);
+    }
+    [self enter:sourceState targetStates:targetStates region:region data:data];
+}
+
+- (void)enter:(TBSMState *)sourceState targetState:(TBSMState *)targetState data:(NSDictionary *)data
 {
     NSUInteger targetLevel = [[targetState.parentNode path] count];
     NSUInteger thisLevel = self.path.count;
@@ -149,12 +165,28 @@
     } else {
         NSArray *targetPath = [targetState.parentNode path];
         id<TBSMNode> node = targetPath[thisLevel];
-        _currentState = node.parentNode;
+        _currentState = (TBSMState *)node.parentNode;
     }
     [self.currentState enter:sourceState targetState:targetState data:data];
 }
 
-- (void)exitState:(TBSMState *)sourceState targetState:(TBSMState *)targetState data:(NSDictionary *)data
+- (void)enter:(TBSMState *)sourceState targetStates:(NSArray *)targetStates region:(TBSMParallelState *)region data:(NSDictionary *)data
+{
+    NSUInteger targetLevel = [[region.parentNode path] count];
+    NSUInteger thisLevel = self.path.count;
+    
+    if (targetLevel == thisLevel) {
+        _currentState = region;
+    } else if (targetLevel > thisLevel) {
+        NSArray *targetPath = [region.parentNode path];
+        id<TBSMNode> node = targetPath[thisLevel];
+        _currentState = (TBSMState *)node.parentNode;
+    }
+    id<TBSMContainingNode> node = (id <TBSMContainingNode>)_currentState;
+    [node enter:sourceState targetStates:targetStates region:region data:data];
+}
+
+- (void)exit:(TBSMState *)sourceState targetState:(TBSMState *)targetState data:(NSDictionary *)data
 {
     [self.currentState exit:sourceState targetState:targetState data:data];
 }
@@ -167,7 +199,7 @@
     TBSMStateMachine *stateMachine = self;
     while (stateMachine) {
         [path insertObject:stateMachine atIndex:0];
-        stateMachine = stateMachine.parentNode.parentNode;
+        stateMachine = (TBSMStateMachine *)stateMachine.parentNode.parentNode;
     }
     return path;
 }